// frontend/xacria-presenze/src/generated-client/model/approvalRequestTab.ts
/**
 * API Presenze-Pro
 * API used by the Presenze-Pro thesis project’s client
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Indica se le richieste che vengono ricercate devono esssere aperte o chiuse
 */
export type ApprovalRequestTab = 'OPEN' | 'CLOSED';

export const ApprovalRequestTab = {
    OPEN: 'OPEN' as ApprovalRequestTab,
    CLOSED: 'CLOSED' as ApprovalRequestTab
};